+++ tests/test_html_entity.py
@@ -57,13 +57,13 @@ class TestHTMLEntity(TreeEqualityTestCase):
         node1 = HTMLEntity("nbsp", named=True, hexadecimal=False)
         node2 = HTMLEntity("107", named=False, hexadecimal=False)
         node3 = HTMLEntity("e9", named=False, hexadecimal=True)
-        for a in (True, False):
-            self.assertEqual("\xa0", node1.__strip__(True, a))
-            self.assertEqual("&nbsp;", node1.__strip__(False, a))
-            self.assertEqual("k", node2.__strip__(True, a))
-            self.assertEqual("&#107;", node2.__strip__(False, a))
-            self.assertEqual("é", node3.__strip__(True, a))
-            self.assertEqual("&#xe9;", node3.__strip__(False, a))
+
+        self.assertEqual("\xa0", node1.__strip__(normalize=True))
+        self.assertEqual("&nbsp;", node1.__strip__(normalize=False))
+        self.assertEqual("k", node2.__strip__(normalize=True))
+        self.assertEqual("&#107;", node2.__strip__(normalize=False))
+        self.assertEqual("é", node3.__strip__(normalize=True))
+        self.assertEqual("&#xe9;", node3.__strip__(normalize=False))
 
     def test_showtree(self):
         """test HTMLEntity.__showtree__()"""